requirements.txt
anyio==4.3.0
asyncpg==0.27.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.89.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
packaging==24.0
pluggy==1.5.0
psycopg2-binary==2.9.6
pydantic==1.10.15
pytest==8.2.1
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
starlette==0.22.0
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.20.0
